still trying
still trying new methods. found a source that lists the steps of d3.js
so going to try to adjust that with my information. realize by looking
through this that I was probably entering my data wrong before (by not
separating it into 2 lines?) so might go back to try to change that if
I get this to work successfully